requirements.txt
#mlflow
fastapi
uvicorn
tensorflow
opencv-python-headless
pillow
python-multipart
dagshub
tensorflow